refactor(politics-detail): modify position-change structures
